Integration 2024 week 42

* !3047 Resolve compilation warnings on ARM
* !3052 Optimize T_ID() macro.
* !3045 CI: Reduce tested DL throughput on Aerial
* !3039 fix for Other SI reception when BWP start != 0
* !2997 websrv regression
* !3041 Update NR SA Tutorials
* !3042 Update TESTBenches.md to reflect current testbench
* !3004 Hotfix to remove an unwanted TX power reduction when using multiple TX antenna ports.
* reduce ssPBCH_BlockPower in gNB config of FHI7.2 pipeline
* !3046 A testsuite and a benchmark for RLC AM mode
* !3038 Various bugfixes for gNB stability, cleanup
* !3051 PUCCH2 DTX detection
* nr_psbchsim: fix segv by providing correct get_nrUE_params()
* !3049 Improve PUSCH and PUCCH SNR stability
* !3054 Allocate and add sD to DRB NSSAI in case sD != 0xffffff

+++ executables/rt_profiling.h
@@ -40,7 +40,7 @@ extern "C"
 // depth of trace in slots
 #define RT_PROF_DEPTH 100
 typedef struct {
-   int absslot_rx[RT_PROF_DEPTH];
+   int64_t absslot_rx[RT_PROF_DEPTH];
    struct timespec return_RU_south_in[RT_PROF_DEPTH];
    struct timespec return_RU_feprx[RT_PROF_DEPTH];
    struct timespec return_RU_prachrx[RT_PROF_DEPTH];
@@ -50,7 +50,7 @@ typedef struct {
 } rt_ru_profiling_t;
 
 typedef struct {
-   int absslot_ux[RT_PROF_DEPTH];
+   int64_t absslot_ux[RT_PROF_DEPTH];
    struct timespec start_L1_RX[RT_PROF_DEPTH];
    struct timespec return_L1_RX[RT_PROF_DEPTH];
    struct timespec start_L1_TX[RT_PROF_DEPTH];
